+++ ft_lst_create.c
@@ -6,7 +6,7 @@
 /*   By: lmeyre <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/02 11:59:08 by lmeyre            #+#    #+#             */
-/*   Updated: 2018/02/18 11:47:23 by lmeyre           ###   ########.fr       */
+/*   Updated: 2018/04/30 09:41:11 by lmeyre           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@ t_list	*ft_create_lst(int size, int content_size)
 	t_list	*curr;
 	t_list	*prev;
 
-	if (!(start = ft_lst_new(content_size)))
+	if (!(start = ft_lst_new((size_t)content_size)))
 		return (NULL);
 	curr = start;
 	prev = NULL;
 	while (size > 1)
 	{
-		if (!(curr->next = ft_lst_new(content_size)))
+		if (!(curr->next = ft_lst_new((size_t)content_size)))
 			return (NULL);
 		curr->prev = prev;
 		prev = curr;
